modules: Move UDisksModuleObject to the daemon
Modules are essentially parts of the daemon and UDisksModuleObject
is an internal implementation. Furthermore modules only need a header
file to implement own objects implementing the UDisksModuleObject
interface.
