+++ modules/vdo/Makefile.am
@@ -38,15 +38,11 @@ libudisks2_vdoincludedir = $(includedir)/udisks2/udisks
 libudisks2_vdoinclude_HEADERS =                                                \
 	$(top_srcdir)/modules/udisksmoduleiface.h                              \
 	$(top_srcdir)/modules/udisksmoduleifacetypes.h                         \
-	$(top_srcdir)/modules/udisksmoduleobject.h                             \
 	$(NULL)
 
 libudisks2_vdo_la_SOURCES =                                                    \
 	$(BUILT_SOURCES)                                                       \
 	$(top_srcdir)/modules/udisksmoduleiface.h                              \
-	$(top_srcdir)/modules/udisksmoduleifacetypes.h                         \
-	$(top_srcdir)/modules/udisksmoduleobject.h                             \
-	$(top_srcdir)/modules/udisksmoduleobject.c                             \
 	udiskslinuxblockvdo.h                                                  \
 	udiskslinuxblockvdo.c                                                  \
 	udiskslinuxmanagervdo.h                                                \
