+++ src/Makefile.am
@@ -101,9 +101,8 @@ libudisks_daemon_la_SOURCES =                                                  \
 	udiskslinuxdevice.h            udiskslinuxdevice.c                     \
 	udisksata.h                    udisksata.c                             \
 	udisksmodulemanager.h          udisksmodulemanager.c                   \
+	udisksmoduleobject.h           udisksmoduleobject.c                    \
 	udisksconfigmanager.h          udisksconfigmanager.c                   \
-	$(top_srcdir)/modules/udisksmoduleobject.h                             \
-	$(top_srcdir)/modules/udisksmoduleobject.c                             \
 	$(BUILT_SOURCES)                                                       \
 	$(NULL)
